fixed norminette and added ft_split

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: dmitrii <dmitrii@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/20 15:25:26 by dmitrypopov       #+#    #+#             */
-/*   Updated: 2024/09/21 18:20:55 by dmitrii          ###   ########.fr       */
+/*   Updated: 2024/09/24 15:54:26 by dmitrii          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *str)
+#include "libft.h"
+
+size_t	ft_strlen(const char *str)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
